+++ platformio/platforms/freescalekinetis.py
@@ -34,7 +34,7 @@ class FreescalekinetisPlatform(BasePlatform):
         },
 
         "framework-mbed": {
-            "default": True
+            "alias": "framework"
         }
     }