+++ platformio/platforms/ststm32.py
@@ -40,26 +40,32 @@ class Ststm32Platform(BasePlatform):
         },
 
         "tool-stlink": {
-            "alias": "uploader",
-            "default": True
+            "alias": "uploader"
         },
 
         "framework-cmsis": {
-            "default": True
+            "alias": "framework"
         },
 
         "framework-spl": {
-            "default": True
+            "alias": "framework"
         },
 
         "framework-libopencm3": {
-            "default": True
+            "alias": "framework"
         },
 
         "framework-mbed": {
-            "default": True
+            "alias": "framework"
         }
     }
 
     def get_name(self):
         return "ST STM32"
+
+    def configure_default_packages(self, envoptions, targets):
+        if envoptions.get("framework") == "cmsis":
+            self.PACKAGES['framework-mbed']['default'] = True
+
+        return BasePlatform.configure_default_packages(
+            self, envoptions, targets)